Created SystemMonitorUI Class to encapsulate the UI functionality
